wip: Factor file stream processors
