+++ src/file-write.ts
@@ -10,9 +10,10 @@ import {
     fsReadAsync
 } from "./util-file"
 import * as fs from 'fs'
-import { Header, getId3Tag } from "./id3-tag"
+import { Header } from "./id3-tag"
 import { WriteOptions } from "./types/write"
 import { hrtime } from "process"
+import { Id3TagRemover } from "./file-stream-processor"
 
 // Must be at least Header.size which is the min size to detect an ID3 header.
 // Naming it help identifying the code handling it.
@@ -92,56 +93,6 @@ function makeTempFilepath(filepath: string) {
     // when running multiple tests in parallel for example.
     // Date.now() resolution is too low.
     return `${filepath}.tmp-${hrtime.bigint()}`
-}
-
-class Id3TagRemover {
-    private buffer: Buffer
-    private rolloverSize = 0
-    continue = false
-
-    constructor(bufferSize: number) {
-        // TODO enforce min buffer size here,
-        // i.e. bufferSize + RolloverBufferSize + 1
-        this.buffer = Buffer.alloc(bufferSize)
-    }
-
-    getReadBuffer() {
-        return this.buffer.subarray(this.rolloverSize)
-    }
-
-    processReadBuffer(readSize: number) {
-        let data = this.buffer.subarray(0, this.rolloverSize + readSize)
-
-        // TODO extract that to id3-tag
-        // Remove tags from `data`
-        let missingBytes = 0
-        const parts = Array.from((function*() {
-            let tag
-            while((tag = getId3Tag(data))) {
-                yield tag.before
-                data = tag.after
-                missingBytes = tag.missingBytes
-            }
-        })())
-
-        // Exclude rollover window on the last part
-        this.rolloverSize = Math.min(RolloverBufferSize, data.length, readSize)
-        const rolloverStart = data.length - this.rolloverSize
-        const rolloverData = Buffer.from(data.subarray(rolloverStart))
-        parts.push(data.subarray(0, rolloverStart))
-
-        const writeBuffer = Buffer.concat(parts)
-
-        // Update rollover window
-        rolloverData.copy(this.buffer)
-
-        this.continue = this.rolloverSize !==0 || missingBytes !== 0
-
-        return {
-            skipBuffer: Buffer.alloc(missingBytes),
-            writeBuffer
-        }
-    }
 }
 
 function copyFileWithoutId3TagSync(
